[backend] test(multi-tenancy): channel API for multi-tenancy (#5718)

+++ openaev-api/src/main/java/io/openaev/rest/channel/ChannelApi.java
@@ -69,7 +69,9 @@ public class ChannelApi extends RestBehavior {
       actionPerformed = Action.READ,
       resourceType = ResourceType.CHANNEL)
   public Channel channel(@PathVariable String channelId) {
-    return channelRepository.findById(channelId).orElseThrow(ElementNotFoundException::new);
+    return channelRepository
+        .findByIdAndTenantId(channelId, TenantContext.getCurrentTenant())
+        .orElseThrow(ElementNotFoundException::new);
   }
 
   @PutMapping({CHANNEL_URI + "/{channelId}", TENANT_CHANNEL_URI + "/{channelId}"})
@@ -81,7 +83,9 @@ public class ChannelApi extends RestBehavior {
   public Channel updateChannel(
       @PathVariable String channelId, @Valid @RequestBody ChannelUpdateInput input) {
     Channel channel =
-        channelRepository.findById(channelId).orElseThrow(ElementNotFoundException::new);
+        channelRepository
+            .findByIdAndTenantId(channelId, TenantContext.getCurrentTenant())
+            .orElseThrow(ElementNotFoundException::new);
     channel.setUpdateAttributes(input);
     channel.setUpdatedAt(Instant.now());
     return channelRepository.save(channel);
@@ -96,7 +100,9 @@ public class ChannelApi extends RestBehavior {
   public Channel updateChannelLogos(
       @PathVariable String channelId, @Valid @RequestBody ChannelUpdateLogoInput input) {
     Channel channel =
-        channelRepository.findById(channelId).orElseThrow(ElementNotFoundException::new);
+        channelRepository
+            .findByIdAndTenantId(channelId, TenantContext.getCurrentTenant())
+            .orElseThrow(ElementNotFoundException::new);
     if (input.getLogoDark() != null) {
       channel.setLogoDark(documentRepository.findById(input.getLogoDark()).orElse(null));
     } else {
@@ -126,6 +132,9 @@ public class ChannelApi extends RestBehavior {
       actionPerformed = Action.DELETE,
       resourceType = ResourceType.CHANNEL)
   public void deleteChannel(@PathVariable String channelId) {
+    if (!channelRepository.existsByIdAndTenantId(channelId, TenantContext.getCurrentTenant())) {
+      throw new ElementNotFoundException();
+    }
     channelRepository.deleteById(channelId);
   }
 
@@ -142,7 +151,9 @@ public class ChannelApi extends RestBehavior {
       @PathVariable String exerciseId, @PathVariable String channelId) {
     ChannelReader channelReader;
     Channel channel =
-        channelRepository.findById(channelId).orElseThrow(ElementNotFoundException::new);
+        channelRepository
+            .findByIdAndTenantId(channelId, TenantContext.getCurrentTenant())
+            .orElseThrow(ElementNotFoundException::new);
 
     Optional<Exercise> exerciseOpt =
         this.exerciseRepository.findByIdAndTenantId(exerciseId, TenantContext.getCurrentTenant());
@@ -203,7 +214,7 @@ public class ChannelApi extends RestBehavior {
     article.setUpdateAttributes(input);
     article.setChannel(
         channelRepository
-            .findById(input.getChannelId())
+            .findByIdAndTenantId(input.getChannelId(), TenantContext.getCurrentTenant())
             .orElseThrow(ElementNotFoundException::new));
     article.setExercise(exercise);
     Article savedArticle = articleRepository.save(article);
@@ -249,7 +260,7 @@ public class ChannelApi extends RestBehavior {
     List<String> currentDocumentIds = article.getDocuments().stream().map(Document::getId).toList();
     article.setChannel(
         channelRepository
-            .findById(input.getChannelId())
+            .findByIdAndTenantId(input.getChannelId(), TenantContext.getCurrentTenant())
             .orElseThrow(ElementNotFoundException::new));
     article.setUpdateAttributes(input);
     // Original List
@@ -313,7 +324,7 @@ public class ChannelApi extends RestBehavior {
     article.setUpdateAttributes(input);
     article.setChannel(
         this.channelRepository
-            .findById(input.getChannelId())
+            .findByIdAndTenantId(input.getChannelId(), TenantContext.getCurrentTenant())
             .orElseThrow(ElementNotFoundException::new));
     article.setScenario(scenario);
     Article savedArticle = this.articleRepository.save(article);
@@ -356,7 +367,7 @@ public class ChannelApi extends RestBehavior {
     List<String> currentDocumentIds = article.getDocuments().stream().map(Document::getId).toList();
     article.setChannel(
         channelRepository
-            .findById(input.getChannelId())
+            .findByIdAndTenantId(input.getChannelId(), TenantContext.getCurrentTenant())
             .orElseThrow(ElementNotFoundException::new));
     article.setUpdateAttributes(input);
     // Original List
